story-book-maker/pyproject.toml
[project]
name = "story-book-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudpickle>=3.1.2",
    "google-adk[eval]>=1.12.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.141.0",
    "google-genai>=1.31.0",
    "greenlet>=3.0.0",
    "litellm>=1.76.0",
    "requests>=2.32.5",
    "sseclient-py>=1.9.0",
]

[dependency-groups]
dev = ["ipykernel>=6.30.1"]
